+++ src/pool-cl/dynamic-fee/interfaces/IPriceFeedTwoOracle.sol
@@ -30,5 +30,5 @@ interface IPriceFeedTwoOracle {
 
     function token1() external view returns (IERC20Metadata);
 
-    function getPriceX96() external view returns (uint160 priceX96);
+    function getPriceX96() external view returns (uint256 priceX96);
 }
